Aitividade_Aula_01

+++ desafio_4.py
@@ -1,6 +1,6 @@
 # DESAFIO 04 - Desenvolva um programa que leia as duas notas de um aluno, calcule e mostra a sua média.
-nota1 = int(input("Digite a nota da primeira prova: "))
-nota2 = int(input("Digite a nota da segunda prova: "))
+nota1 = float(input("Digite a nota da primeira prova: "))
+nota2 = float(input("Digite a nota da segunda prova: "))
 
 nota_soma = nota1 + nota2
 nota_media = nota_soma / 2
